+++ src/ecp/common/generator/ecp_g_jarosz.cc
@@ -3179,7 +3179,7 @@ void elipsoid::ecp_save_trajectory()
 		e = errno;
 		perror("ECP: Send() to UI failed");
 		ecp_t.sr_ecp_msg->message(lib::SYSTEM_ERROR, e, "ECP: Send() to UI failed");
-		throw generator::ECP_error(lib::SYSTEM_ERROR, (uint64_t) 0);
+		throw generator::ECP_error(lib::SYSTEM_ERROR, 0);
 	}
 
 	if (ui_to_ecp_rep.reply == lib::QUIT) // Nie wybrano nazwy pliku lub zrezygnowano z zapisu
